move contents of __init__.py to api.py
